--- test-results/.last-run.json
@@ -1,6 +1,0 @@
-{
-  "status": "failed",
-  "failedTests": [
-    "6b6dc33e290e5a0a43b1-da054245900576bf1b5a"
-  ]
-}
